Added custom strategies for users.
Now a user can describe an interest, an investment idea or a strategy and we will carry out analysis based on this. In this front end we create new and manage custom user specific strategies.
In the backend we locate interesting topics, grab relevant information, write a custom analysis which can then be loaded by the frontend and interacted with by the user.